Move Debug into its own namespace
